story-generator-agentic-rag/requirements.txt
streamlit==1.37.1
chromadb==0.5.5
google-cloud-aiplatform==1.66.0
vertexai==1.66.0
python-dotenv==1.0.1
PyPDF2==3.0.1
gitpython==3.1.43
beautifulsoup4==4.12.3
requests==2.32.3
orjson==3.10.7
pydantic==2.8.2
tenacity==8.5.0
uvloop==0.19.0; platform_system != "Windows"
